+++ sen/sinfo.go
@@ -91,7 +91,7 @@ func buildTagFields(rt reflect.Type, out, pretty, embedded, omitEmpty bool) (fa 
 		if len(name) == 0 || 'a' <= name[0] || name[0] == '_' {
 			continue
 		}
-		if f.Anonymous && !out {
+		if f.Anonymous && !out && f.Type.Kind() != reflect.Interface {
 			if f.Type.Kind() == reflect.Ptr {
 				for _, fi := range buildTagFields(f.Type.Elem(), out, pretty, embedded, omitEmpty) {
 					fi.index = append([]int{i}, fi.index...)
@@ -107,6 +107,7 @@ func buildTagFields(rt reflect.Type, out, pretty, embedded, omitEmpty bool) (fa 
 			}
 		} else {
 			asString := false
+			omit := omitEmpty
 			key := f.Name
 			if tag, ok := f.Tag.Lookup("json"); ok && 0 < len(tag) {
 				parts := strings.Split(tag, ",")
@@ -125,13 +126,13 @@ func buildTagFields(rt reflect.Type, out, pretty, embedded, omitEmpty bool) (fa 
 				for _, p := range parts[1:] {
 					switch p {
 					case "omitempty":
-						omitEmpty = true
+						omit = true
 					case "string":
 						asString = true
 					}
 				}
 			}
-			fa = append(fa, newFinfo(&f, key, omitEmpty, asString, pretty, embedded))
+			fa = append(fa, newFinfo(&f, key, omit, asString, pretty, embedded))
 		}
 	}
 	return
@@ -144,7 +145,7 @@ func buildExactFields(rt reflect.Type, out, pretty, embedded, omitEmpty bool) (f
 		if len(name) == 0 || 'a' <= name[0] || name[0] == '_' {
 			continue
 		}
-		if f.Anonymous && !out {
+		if f.Anonymous && !out && f.Type.Kind() != reflect.Interface {
 			if f.Type.Kind() == reflect.Ptr {
 				for _, fi := range buildExactFields(f.Type.Elem(), out, pretty, embedded, omitEmpty) {
 					fi.index = append([]int{i}, fi.index...)
@@ -172,7 +173,7 @@ func buildLowFields(rt reflect.Type, out, pretty, embedded, omitEmpty bool) (fa 
 		if len(name) == 0 || 'a' <= name[0] || name[0] == '_' {
 			continue
 		}
-		if f.Anonymous && !out {
+		if f.Anonymous && !out && f.Type.Kind() != reflect.Interface {
 			if f.Type.Kind() == reflect.Ptr {
 				for _, fi := range buildLowFields(f.Type.Elem(), out, pretty, embedded, omitEmpty) {
 					fi.index = append([]int{i}, fi.index...)