+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 The structure and content of this file follows [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.26.7] - unreleased
+### Fixed
+- Fixed then bleed over of omitempty in struct tags. Issue #205.
+- Fixed encoding anonymous interface in structs. Issue #115
+
 ## [1.26.6] - 2025-05-27
 ### Fixed
 - Types derived from core type such as `type Quux int` now alt.Decompose correctly.